% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{aggregate_users}
\alias{aggregate_users}
\title{aggregate_users}
\usage{
aggregate_users(trip_data)
}
\arguments{
\item{trip_data}{The one-row-per-trip data (as output by `aggregate_trips`).}
}
\value{
A tibble of all available user data.
}
\description{
Aggregate user data.
}
\details{
Create a one-row-per-user dataset from the output of `aggregate_trips`.
}
\examples{
\dontrun{
full_data <- import_full()
trip_data <- aggregate_trips(full_data)
user_data <- aggregate_trips(trip_data)
}

}
