% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{byMonth}
\alias{byMonth}
\alias{june2018}
\alias{july2018}
\alias{august2018}
\alias{september2018}
\alias{october2018}
\alias{november2018}
\alias{april2019}
\alias{may2019}
\alias{june2019}
\alias{july2019}
\alias{august2019}
\alias{september2019}
\alias{october2019}
\alias{november2019}
\alias{june2020}
\alias{july2020}
\alias{august2020}
\alias{september2020}
\alias{october2020}
\title{By-month trajectory data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36773 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2054773 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5802790 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8927495 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7331158 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2598266 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4681751 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3379888 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4875254 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4369828 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4006793 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3360060 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2223486 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 879494 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 435629 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2641636 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2890749 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2681825 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 361804 rows and 6 columns.
}
\usage{
june2018

july2018

august2018

september2018

october2018

november2018

april2019

may2019

june2019

july2019

august2019

september2019

october2019

november2019

june2020

july2020

august2020

september2020

october2020
}
\description{
The by-month datasets contain monthly trajectory data for all the months that ValleyBike has
been in active operation (normally April-November each year). The point data (latitude, longitude)
was collected during every trip, at 5-second intervals.
}
\section{Variables}{

\itemize{
 \item route_id (character), the trip's unique route id (primary key)
 \item user_id (character), the rider's unique user id
 \item bike (character), unique bike id
 \item time (datetime), the time at which the location was recorded (down to seconds)
 \item longitude (double), the longitude of the bike at that point in time
 \item latitude (double), the latitude of the bike at that point in time
}
}

\keyword{datasets}
