% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{download_data}
\alias{download_data}
\title{download_data}
\usage{
download_data(path, overwrite = FALSE)
}
\arguments{
\item{path}{The path where to download the data files. Presumably, this will be inst/extdata.}

\item{overwrite}{Whether to overwrite the existing files at the destination path. Defaults to FALSE.}
}
\description{
Download raw data files
}
\details{
Download all available .csv.gz raw trajectory data files for the years 2018-2020 into a specified
directory. Intended usage is for updating the files in inst/extdata to mirror those online.
}
\examples{
\dontrun{
download_data(path = "inst/extdata", overwrite = TRUE)
}

}
