% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_files}
\alias{download_files}
\title{download_files}
\usage{
download_files(path, overwrite = FALSE)
}
\arguments{
\item{path}{The path where to download the data files.}

\item{overwrite}{Whether to overwrite the existing files at the given path. Defaults to \code{FALSE}.}
}
\description{
Download raw data files
}
\details{
Download all available .csv.gz raw trajectory data files for the years 2018-2020 into a specified
directory.
}
\examples{
\dontrun{
download_files(path = "~/Desktop/raw-data")
}

}
