% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_distance}
\alias{get_distance}
\title{get_distance}
\usage{
get_distance(route_id)
}
\arguments{
\item{route_id}{The route ID (as a string)}
}
\value{
The distance of the route ID's corresponding trip (as a numeric constant)
}
\description{
Get distance of trip
}
\details{
Get the distance of some trip, based on its route ID.
}
\examples{
get_distance("route_04_2019@01253ad9-cfb8-44a5-9218-bff8109e4880")

}
