% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_duration}
\alias{get_duration}
\title{get_duration}
\usage{
get_duration(route_id)
}
\arguments{
\item{route_id}{The route ID (as a string)}
}
\value{
The duration of the route ID's corresponding trip (as a numeric constant)
}
\description{
Get duration of trip
}
\details{
Get the duration of some trip, based on its route ID.
}
\examples{
get_duration("route_04_2019@01253ad9-cfb8-44a5-9218-bff8109e4880")

}
