% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{get_full_data}
\alias{get_full_data}
\title{get_full_data}
\usage{
get_full_data()
}
\value{
A 65,975,278 x 6 tibble of all available trajectory data.
}
\description{
Get the full trajectory data (raw)
}
\details{
Get all available trajectory data for the years 2018-2020, in raw format.
}
\examples{
\dontrun{
full_data <- get_full_data()
}

}
