% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stops}
\alias{get_stops}
\title{get_stops}
\usage{
get_stops(route_id)
}
\arguments{
\item{route_id}{The route ID (as a string)}
}
\value{
A data frame of all stops in the given trip (with latitude, longitude, and time)
}
\description{
Get stops in a trip
}
\details{
Get the stops in a trip, based on its route ID.
}
\examples{
get_stops("route_04_2019@01253ad9-cfb8-44a5-9218-bff8109e4880")

}
