% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_trajectory_data}
\alias{get_trajectory_data}
\title{get_trajectory_data}
\usage{
get_trajectory_data(
  start_time = "2018-06-01 00:00:00",
  end_time = "2019-11-30 23:59:59",
  route_id,
  user_id,
  bike
)
}
\arguments{
\item{start_time}{The start time of the queried range (as a string of the form "YYYY-MM-DD HH:MM:SS")}

\item{end_time}{The end time of the queried range (as a string of the form "YYYY-MM-DD HH:MM:SS")}

\item{route_id}{A route ID to filter by (as a string)}

\item{user_id}{A user ID to filter by (as a string)}

\item{bike}{A bike number to filter by (as a string)}
}
\value{
A data frame of trajectory data, filtered according to the user's specifications.
}
\description{
Get trajectory data
}
\details{
Get trajectory data for a specific timerange, route ID, user_id, or bike. Arguments can
be mixed and matched. Not specifying any arguments will return all of the 2018-2019
trajectory data (400 Mb).
}
\examples{
data <- get_trajectory_data(start_time = "2018-06-01 00:00:00", end_time = "2018-06-29 05:04:44",
                            user_id = "72491657-3115-4497-879d-000830f2352b")
}
