% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_day}
\alias{import_day}
\title{import_day}
\usage{
import_day(day, return = c("all", "clean", "anomalous"))
}
\arguments{
\item{day}{The day for which to import the data (as a string of the form "YYYY-MM-DD")}

\item{return}{The type of data to return (one of "all", "clean", "anomalous"). Defaults to "all"}
}
\value{
A data frame of available trajectory data for that specific day.
}
\description{
Import trajectory data for one day.
}
\details{
Import trajectory data for a specific day. The user can choose to import the raw data, 
the clean data (i.e. the raw data minus any anomalous observations), or the anomalous data.
}
\examples{
data_22_may_2019 <- import_day("2019-05-22", return = "clean")

}
