% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_full}
\alias{import_full}
\title{import_full}
\usage{
import_full()
}
\value{
A 65,975,278 x 6 tibble of all available trajectory data.
}
\description{
Import full trajectory data (raw)
}
\details{
Import all available trajectory data for the years 2018-2020, in raw format.
}
\examples{
\dontrun{
full_data <- import_full()
}

}
