% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_month}
\alias{import_month}
\title{import_month}
\usage{
import_month(month, ...)
}
\arguments{
\item{month}{The month for which to import the data (as a string of the form "YYYY-MM").}

\item{...}{Further parameters to pass to `import_day()` (e.g. `return` or `future_cutoff`).}
}
\value{
A tibble of available trajectory data for that specific month.
}
\description{
Import trajectory data for one month.
}
\details{
Import trajectory data for a specific month. The user can choose to import the raw data,
the clean data (i.e. the raw data minus any anomalous observations), or the anomalous data.
}
\examples{
\dontrun{
june2018 <- import_month("2018-06")
}

}
