% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly}
\alias{monthly}
\alias{june2018}
\alias{july2018}
\alias{august2018}
\alias{september2018}
\alias{october2018}
\alias{november2018}
\alias{april2019}
\alias{may2019}
\alias{june2019}
\alias{july2019}
\alias{august2019}
\alias{september2019}
\alias{october2019}
\alias{november2019}
\alias{june2020}
\alias{july2020}
\alias{august2020}
\alias{september2020}
\alias{october2020}
\alias{november2020}
\alias{december2020}
\alias{january2021}
\alias{february2021}
\alias{march2021}
\alias{april2021}
\alias{may2021}
\title{Monthly trajectory data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36773 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2054773 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5802790 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8927495 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7331158 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2598266 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4681751 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3379888 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4875254 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4369828 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4006793 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3360060 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2223486 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 879494 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 435629 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2641636 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2890749 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2681825 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 361804 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1013172 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 363270 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 296398 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140010 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1605535 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2349597 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3622175 rows and 6 columns.
}
\usage{
data("june2018")

data("july2018")

data("august2018")

data("september2018")

data("october2018")

data("november2018")

data("april2019")

data("may2019")

data("june2019")

data("july2019")

data("august2019")

data("june2018")

data("october2019")

data("november2019")

data("june2020")

data("july2020")

data("august2020")

data("september2020")

data("october2020")

data("november2020")

data("december2020")

data("january2021")

data("february2021")

data("march2021")

data("april2021")

data("may2021")
}
\description{
The monthly datasets contain month-by-month trajectory data for all the months that ValleyBike has
been in active operation. The point data (latitude, longitude)
was collected during every trip, at 5-second intervals.
}
\section{Variables}{

\itemize{
 \item route_id (character), the trip's unique route id (primary key)
 \item user_id (character), the rider's unique user id
 \item bike (character), unique bike id
 \item time (datetime), the time at which the location was recorded (down to seconds)
 \item longitude (double), the longitude of the bike at that point in time
 \item latitude (double), the latitude of the bike at that point in time
}
}

\keyword{datasets}
