% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{users}
\alias{users}
\title{ValleyBike user statistics over 2018-2019}
\format{
A 12,553 x 10 data frame
}
\usage{
users
}
\description{
This data set is contains anonymous statistics for ValleyBike users in
2018, 2019, and 2020.
}
\section{Variables}{

\itemize{
 \item user_id <chr>, the user's unique id (primary key)
 \item num_trips <int>, the total number of trips taken by the user
 \item first_trip <dttm> the date-time of the user's first recorded trip
 \item last_trip <dttm> the date-time of the user's last recorded trip
 \item mean_trip_duration <dbl>, the user's mean trip duration
 \item median_trip_duration <dbl>, the user's median trip duration
 \item most_freq_start_station <chr>, the station at which the user most
       frequently starts a trip
 \item num_starting_there <int>, the number of trips starting at the user's
       most frequent start station
 \item most_freq_end_station <chr>, the station at which the user most
       frequently ends a trip
 \item num_ending_there <int>, the number of trips ending at the user's
       most frequent end station
}
}

\keyword{datasets}
