% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{aggregate_trips}
\alias{aggregate_trips}
\title{aggregate_trips}
\usage{
aggregate_trips(full_data)
}
\arguments{
\item{full_data}{The full trajectory data (as output by `import_full`).}
}
\value{
A tibble of all available trip data.
}
\description{
Aggregate trip data.
}
\details{
Create a one-row-per-trip dataset from the output of `import_full`.
}
\examples{
\dontrun{
full_data <- import_full()
trips <- aggregate_trips(full_data)
}

}
