% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{byMonth}
\alias{byMonth}
\alias{june2018}
\alias{july2018}
\alias{august2018}
\alias{september2018}
\alias{october2018}
\alias{november2018}
\alias{april2019}
\alias{may2019}
\alias{june2019}
\alias{july2019}
\alias{august2019}
\alias{september2019}
\alias{october2019}
\alias{november2019}
\title{By-month trajectory data}
\format{
An object of class \code{data.frame} with 23516 rows and 6 columns.

An object of class \code{data.frame} with 2049390 rows and 6 columns.

An object of class \code{data.frame} with 5835718 rows and 6 columns.

An object of class \code{data.frame} with 7746958 rows and 6 columns.

An object of class \code{data.frame} with 7456798 rows and 6 columns.

An object of class \code{data.frame} with 2668702 rows and 6 columns.

An object of class \code{data.frame} with 4636234 rows and 6 columns.

An object of class \code{data.frame} with 3380879 rows and 6 columns.

An object of class \code{data.frame} with 4979741 rows and 6 columns.

An object of class \code{data.frame} with 4454943 rows and 6 columns.

An object of class \code{data.frame} with 4177156 rows and 6 columns.

An object of class \code{data.frame} with 3406582 rows and 6 columns.

An object of class \code{data.frame} with 2255826 rows and 6 columns.

An object of class \code{data.frame} with 878730 rows and 6 columns.
}
\usage{
june2018

july2018

august2018

september2018

october2018

november2018

april2019

may2019

june2019

july2019

august2019

september2019

october2019

november2019
}
\description{
The by-month data sets (for 2018-2019) contain monthly trajectory data
(latitude, longitude) collected during every trip, at 5-second intervals.
These datasets are quite large (a few million entries), so they might lag
your R session.
}
\section{Variables}{

\itemize{
 \item route_id <chr>, the trip's unique route id (primary key)
 \item user_id <chr>, the rider's unique user id
 \item bike <chr>, unique bike id
 \item time <dttm>, the time at which the location was recorded (down to seconds)
 \item longitude <dbl>, the longitude of the bike at that point in time
 \item latitude <dbl>, the latitude of the bike at that point in time
}
}

\keyword{datasets}
