% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-consistency.R
\name{import_full}
\alias{import_full}
\title{import_full}
\usage{
import_full()
}
\value{
A 60,910,226 x 6 data frame of all available trajectory data, in raw format.
}
\description{
Import full trajectory data (raw)
}
\details{
Get trajectory data for a specific timerange, route ID, user_id, or bike. Arguments can
be mixed and matched. Not specifying any arguments will return all of the 2018-2019
trajectory data (400 Mb).
}
\examples{
\dontrun{
full_data <- import_full()
}

}
