% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trips}
\alias{trips}
\title{ValleyBike trips over 2018-2020}
\format{
A tibble
}
\usage{
trips
}
\description{
This data set is an aggregated one-row-per-trip version of the original
point-in-time ValleyBike data for the years 2018, 2019, and 2020.
}
\section{Variables}{

\itemize{
 \item route_id (character), the trip's unique route id (primary key)
 \item user_id (character), the rider's unique user id
 \item bike (character), unique bike id
 \item start_time (datetime), the trip's starting date-time (EDT)
 \item end_time (datetime), the trip's ending date-time (EDT)
 \item start_station (character), the trip's starting station
 \item start_latitude (double), the trip's starting latitude
 \item start_longitude (double), the trip's starting longitude
 \item end_station (character), the trip's ending station
 \item end_latitude (double), the trip's ending latitude
 \item end_longitude (double), the trip's ending longitude
 \item duration (double), the trip's duration (in seconds)
}
}

\keyword{datasets}
