% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{users}
\alias{users}
\title{ValleyBike user statistics over 2018-2019}
\format{
A tibble
}
\usage{
users
}
\description{
This dataset contains anonymous statistics for ValleyBike users in 2018, 2019, 2020, and 2021.
}
\section{Variables}{

\itemize{
 \item user_id (character), the user's unique id (primary key)
 \item trips (integer), the total number of trips taken by the user
 \item min_trip_duration (double), the user's minimum trip duration
 \item mean_trip_duration (double), the user's mean trip duration
 \item median_trip_duration (double), the user's median trip duration
 \item max_trip_duration (double), the user's maximum trip duration
 \item first_trip_time (datetime), the datetime of the user's first recorded trip
 \item last_trip_time (datetime), the datetime of the user's last recorded trip
 \item top_start_station (character), the station at which the user most frequently starts a trip
 \item top_start_station_trips (integer), the number of trips starting at the top start station
 \item top_end_station (character), the station at which the user most frequently ends a trip
 \item top_end_station_trips (integer), the number of trips ending at the top end station
}
}

\keyword{datasets}
